formatDate関数の共通化
