+++ src/pages/TopPage/components/PostCard/PostCard.jsx
@@ -1,16 +1,9 @@
 import { Link } from "react-router-dom";
+import formatDate from "../../../../utils/formatDate";
+
 import classes from "./PostCard.module.css";
 
 const PostCard = ({ post }) => {
-  // 日付を取得する関数
-  const formatDate = (isoString, separator) => {
-    const date = new Date(isoString);
-    const year = date.getFullYear();
-    const month = date.getMonth() + 1;
-    const day = date.getDate();
-    return `${year}${separator}${month}${separator}${day}`;
-  };
-
   // 記事の本文を抜粋する関数
   const getPreviewContent = (html, maxLines = 2) => {
     const lines = html.split('<br/>');
