+++ src/utils/formatDate.js
@@ -1,0 +1,9 @@
+const formatDate = (isoString, separator) => {
+  const date = new Date(isoString);
+  const year = date.getFullYear();
+  const month = date.getMonth() + 1;
+  const day = date.getDate();
+  return `${year}${separator}${month}${separator}${day}`;
+};
+
+export default formatDate;
